first working code for simple LCA resolver
